+++ source/internal_extensions/omni.cuopt.service/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) .
 
+## [1.3.1] - 2025-11-02
+### Changed
+- Make omni.pip.cloud an explicit dependency for requests module usage
+
 ## [1.3.0] - 2025-06-30
 ### Fixed
 - Updated access to cuOpt OSS service