+++ source/internal_extensions/omni.cuopt.visualization/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) .
 
+## [1.3.2] - 2025-11-02
+### Changed
+- Cleanup unused imports
+
 ## [1.3.1] - 2025-10-27
 ### Changed
 - Make omni.isaac.ml_archive an explicit test dependency